+++ moshi-kotlin-codegen/src/main/java/com/squareup/moshi/kotlin/codegen/api/TargetType.kt
@@ -28,6 +28,8 @@ public data class TargetType(
   val typeVariables: List<TypeVariableName>,
   val isDataClass: Boolean,
   val visibility: KModifier,
+  val isValueClass: Boolean,
+  val isInline: Boolean = false,
 ) {
 
   init {