+++ moshi/src/main/java/com/squareup/moshi/JsonAdapter.kt
@@ -43,7 +43,7 @@ public abstract class JsonAdapter<T> {
    */
   @CheckReturnValue
   @Throws(IOException::class)
-  public abstract fun fromJson(reader: JsonReader): T?
+  public abstract fun fromJson(reader: JsonReader): T
 
   /**
    * Decodes a nullable instance of type [T] from the given [source].
@@ -53,7 +53,7 @@ public abstract class JsonAdapter<T> {
    */
   @CheckReturnValue
   @Throws(IOException::class)
-  public fun fromJson(source: BufferedSource): T? = fromJson(JsonReader.of(source))
+  public fun fromJson(source: BufferedSource): T = fromJson(JsonReader.of(source))
 
   /**
    * Decodes a nullable instance of type [T] from the given `string`.
@@ -63,7 +63,7 @@ public abstract class JsonAdapter<T> {
    */
   @CheckReturnValue
   @Throws(IOException::class)
-  public fun fromJson(@Language("JSON") string: String): T? {
+  public fun fromJson(@Language("JSON") string: String): T {
     val reader = JsonReader.of(Buffer().writeUtf8(string))
     val result = fromJson(reader)
     if (!isLenient && reader.peek() != JsonReader.Token.END_DOCUMENT) {
@@ -74,17 +74,17 @@ public abstract class JsonAdapter<T> {
 
   /** Encodes the given [value] with the given [writer]. */
   @Throws(IOException::class)
-  public abstract fun toJson(writer: JsonWriter, value: T?)
+  public abstract fun toJson(writer: JsonWriter, value: T)
 
   @Throws(IOException::class)
-  public fun toJson(sink: BufferedSink, value: T?) {
+  public fun toJson(sink: BufferedSink, value: T) {
     val writer = JsonWriter.of(sink)
     toJson(writer, value)
   }
 
   /** Encodes the given [value] into a String and returns it. */
   @CheckReturnValue
-  public fun toJson(value: T?): String {
+  public fun toJson(value: T): String {
     val buffer = Buffer()
     try {
       toJson(buffer, value)
@@ -104,7 +104,7 @@ public abstract class JsonAdapter<T> {
    * and as a [java.math.BigDecimal] for all other types.
    */
   @CheckReturnValue
-  public fun toJsonValue(value: T?): Any? {
+  public fun toJsonValue(value: T): Any? {
     val writer = `-JsonValueWriter`()
     return try {
       toJson(writer, value)
@@ -119,7 +119,7 @@ public abstract class JsonAdapter<T> {
    * strings, numbers, booleans and nulls.
    */
   @CheckReturnValue
-  public fun fromJsonValue(value: Any?): T? {
+  public fun fromJsonValue(value: Any?): T {
     val reader = `-JsonValueReader`(value)
     return try {
       fromJson(reader)
@@ -138,7 +138,7 @@ public abstract class JsonAdapter<T> {
     return object : JsonAdapter<T>() {
       override fun fromJson(reader: JsonReader) = delegate.fromJson(reader)
 
-      override fun toJson(writer: JsonWriter, value: T?) {
+      override fun toJson(writer: JsonWriter, value: T) {
         val serializeNulls = writer.serializeNulls
         writer.serializeNulls = true
         try {
@@ -160,9 +160,10 @@ public abstract class JsonAdapter<T> {
    * nulls.
    */
   @CheckReturnValue
-  public fun nullSafe(): JsonAdapter<T> {
+  public fun nullSafe(): JsonAdapter<T?> {
+    @Suppress("UNCHECKED_CAST")
     return when (this) {
-      is NullSafeJsonAdapter<*> -> this
+      is NullSafeJsonAdapter<*> -> this as JsonAdapter<T?>
       else -> NullSafeJsonAdapter(this)
     }
   }
@@ -175,11 +176,26 @@ public abstract class JsonAdapter<T> {
    * handled elsewhere. This should only be used to fail on explicit nulls.
    */
   @CheckReturnValue
-  public fun nonNull(): JsonAdapter<T> {
-    return when (this) {
-      is NonNullJsonAdapter<*> -> this
+  public fun nonNull(): JsonAdapter<T & Any> {
+    val finalizedAdapter = when (this) {
+      is NonNullJsonAdapter<*> -> {
+        @Suppress("UNCHECKED_CAST")
+        this as JsonAdapter<T?>
+      }
+
       else -> NonNullJsonAdapter(this)
     }
+
+    return finalizedAdapter.unsafeNonNull()
+  }
+
+  /**
+   * Internal cast-only nonNull for adapters that are known to be internally null-safe but don't
+   * need to allow null inputs.
+   */
+  internal fun unsafeNonNull(): JsonAdapter<T & Any> {
+    @Suppress("UNCHECKED_CAST")
+    return this as JsonAdapter<T & Any>
   }
 
   /** Returns a JSON adapter equal to this, but is lenient when reading and writing. */
@@ -187,7 +203,7 @@ public abstract class JsonAdapter<T> {
   public fun lenient(): JsonAdapter<T> {
     val delegate: JsonAdapter<T> = this
     return object : JsonAdapter<T>() {
-      override fun fromJson(reader: JsonReader): T? {
+      override fun fromJson(reader: JsonReader): T {
         val lenient = reader.lenient
         reader.lenient = true
         return try {
@@ -197,7 +213,7 @@ public abstract class JsonAdapter<T> {
         }
       }
 
-      override fun toJson(writer: JsonWriter, value: T?) {
+      override fun toJson(writer: JsonWriter, value: T) {
         val lenient = writer.isLenient
         writer.isLenient = true
         try {
@@ -216,7 +232,7 @@ public abstract class JsonAdapter<T> {
 
   /**
    * Returns a JSON adapter equal to this, but that throws a [JsonDataException] when
-   * [unknown names and values][JsonReader.setFailOnUnknown] are encountered.
+   * [unknown names and values][JsonReader.failOnUnknown] are encountered.
    * This constraint applies to both the top-level message handled by this type adapter as well as
    * to nested messages.
    */
@@ -224,7 +240,7 @@ public abstract class JsonAdapter<T> {
   public fun failOnUnknown(): JsonAdapter<T> {
     val delegate: JsonAdapter<T> = this
     return object : JsonAdapter<T>() {
-      override fun fromJson(reader: JsonReader): T? {
+      override fun fromJson(reader: JsonReader): T {
         val skipForbidden = reader.failOnUnknown
         reader.failOnUnknown = true
         return try {
@@ -234,7 +250,7 @@ public abstract class JsonAdapter<T> {
         }
       }
 
-      override fun toJson(writer: JsonWriter, value: T?) {
+      override fun toJson(writer: JsonWriter, value: T) {
         delegate.toJson(writer, value)
       }
 
@@ -257,11 +273,11 @@ public abstract class JsonAdapter<T> {
   public fun indent(indent: String): JsonAdapter<T> {
     val delegate: JsonAdapter<T> = this
     return object : JsonAdapter<T>() {
-      override fun fromJson(reader: JsonReader): T? {
+      override fun fromJson(reader: JsonReader): T {
         return delegate.fromJson(reader)
       }
 
-      override fun toJson(writer: JsonWriter, value: T?) {
+      override fun toJson(writer: JsonWriter, value: T) {
         val originalIndent = writer.indent
         writer.indent = indent
         try {